Refactor to make it easier to profile.
