+++ Brocfile.js
@@ -2,7 +2,9 @@
 
 var EmberApp = require('ember-cli/lib/broccoli/ember-app');
 
-var app = new EmberApp();
+var app = new EmberApp({
+  minifyJS: { enabled: false }
+});
 
 // Use `app.import` to add additional libraries to the generated
 // output files.
